project-bolt-sb1-1xjn4d8a/project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.10.0.82
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.3.0
tinydb==4.8.0
reportlab==4.2.0
aiofiles==23.2.1
python-dotenv==1.0.1
email-validator==2.2.0
reportlab==4.2.0
celery==5.3.0
redis==4.5.0
python-crontab==2.7.1
pydantic-settings==2.0.0
python-dateutil==2.8.2
validators==0.20.0